Move selectionActionType to store/config/app
